+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## vNEXT
+- Generate Solidity documentation. (#40)
 - Add Bellecour poco address to config and harmonize deployment artifacts. (#39)
 - Revert always explicit message on failed transfers. (#38)
 - Add configuration for deployments on custom test networks. (#37)